% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_trim}
\alias{df_trim}
\title{Trim Dataframe Rows with NA Values}
\usage{
df_trim(x)
}
\arguments{
\item{x}{the data.frame to trim.}
}
\value{
The data.frame 'x' with rows containing NA values removed.
}
\description{
Trim rows containing NA values from a 'data.frame' object. A performant
    version of na.omit.data.frame with no data validation or checking.
}
\examples{
data <- data.frame(c(1:4, NA), c(NA, 2:5))
df_trim(data)

}
