% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku.R
\name{drawGuide}
\alias{drawGuide}
\title{drawGuide}
\usage{
drawGuide(label, left, top)
}
\arguments{
\item{label}{a function returning the character string to be shown.}

\item{left}{the horizontal position of the guide in pixels.}

\item{top}{the vertical position of the guide in pixels.}
}
\value{
An object of class 'shiny.tag' comprising the HTML to be rendered.
}
\description{
Internal function used by ichimoku to draw the axis guides for interactive
    Shiny plots.
}
\keyword{internal}
