% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku.R
\name{drawInfotip}
\alias{drawInfotip}
\title{drawInfotip}
\usage{
drawInfotip(sdata, left_px, top_px)
}
\arguments{
\item{sdata}{the selected data frame row.}

\item{left_px}{the horizontal cursor position in pixels.}

\item{top_px}{the vertical cursor position in pixels.}
}
\value{
An object of class 'shiny.tag' comprising the HTML to be rendered.
}
\description{
Internal function used by ichimoku to draw the infotip for interactive Shiny
    plots.
}
\keyword{internal}
