% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{hasStrat}
\alias{hasStrat}
\title{hasStrat}
\usage{
hasStrat(x)
}
\arguments{
\item{x}{an object.}
}
\value{
A logical value of TRUE if the 'strat' attribute of 'x' is set,
    otherwise FALSE.
}
\description{
A function for checking if an object contains a strategy.
}
\details{
Designed to be used by ichimoku functions that are either S3
    methods for class 'ichimoku' or after validation that 'x' is an ichimoku
    object, hence there is no check on the class of 'x' within this function.
}
\examples{
cloud <- ichimoku(sample_ohlc_data)
strat <- strat(cloud)

# TRUE:
hasStrat(strat)
# FALSE:
hasStrat(cloud)

}
