% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ichimoku_df}
\alias{ichimoku_df}
\title{Convert ichimoku to data.frame}
\usage{
ichimoku_df(x, keep.attrs)
}
\arguments{
\item{x}{an object of class 'ichimoku'.}

\item{keep.attrs}{(optional) if set to TRUE, will preserve any custom
attributes set on the original object.}
}
\value{
A 'data.frame' object. The ichimoku object index is preserved as the
    first column with header 'index'.
}
\description{
A performant 'ichimoku' to 'data.frame' constructor with no data validation
    or checking.
}
\examples{
cloud <- ichimoku(sample_ohlc_data)
df <- ichimoku_df(cloud)
str(df)

df2 <- ichimoku_df(cloud, keep.attrs = TRUE)
str(df2)

}
