% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{ichimoku_read}
\alias{ichimoku_read}
\title{Read ichimoku objects from Arrow Archive}
\usage{
ichimoku_read(file)
}
\arguments{
\item{file}{A character file name or URI, raw vector, an Arrow input stream,
or a FileSystem with path (SubTreeFileSystem). If a file name or URI, an
Arrow InputStream will be opened and closed when finished. If an input
stream is provided, it will be left open.}
}
\value{
The ichimoku object that was originally archived.
}
\description{
Used to read ichimoku objects from archive storage in the Apache Arrow IPC
    file format.
}
\details{
Used to read ichimoku objects archived using the
    \code{\link{ichimoku_write}} function.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

filename <- tempfile()
ichimoku_write(cloud, filename)

restored <- ichimoku_read(filename)
all.equal(cloud, restored, check.attributes = FALSE)
unlink(filename)

}
