% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{ichimoku_write}
\alias{ichimoku_write}
\title{Write ichimoku objects to Arrow Archive}
\usage{
ichimoku_write(object, file)
}
\arguments{
\item{object}{an ichimoku object.}

\item{file}{A string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
Invisible NULL. 'object' is written to 'file' as a side effect.
}
\description{
Used to write ichimoku objects to archive storage in the Apache Arrow IPC
    file format.
}
\details{
If the file write operation has been successful, a confirmation will
    be printed to the console. Use \code{\link{ichimoku_read}} to read
    ichimoku objects archived using this function.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

filename <- tempfile()
ichimoku_write(cloud, filename)

restored <- ichimoku_read(filename)
all.equal(cloud, restored, check.attributes = FALSE)
unlink(filename)

}
