% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxOver}
\alias{maxOver}
\title{Maximum Over a Rolling Window}
\usage{
maxOver(x, window)
}
\arguments{
\item{x}{a vector.}

\item{window}{size of the rolling window.}
}
\value{
A vector of the same length as 'x' with elements 1 to
    (length(window) - 1) containing NAs.
}
\description{
Used by \link{ichimoku} to calculate the maximum over a rolling window.
}
\details{
Fast implementation with no error checking or NA handling.
}
\examples{
maxOver(sample_ohlc_data$close[1:10], 3L)

}
