% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minOver}
\alias{minOver}
\title{Minimum over a rolling window}
\usage{
minOver(x, window)
}
\arguments{
\item{x}{a vector.}

\item{window}{size of the rolling window.}
}
\value{
A vector of the same length as 'x' with elements 1 to
    (length(window) - 1) containing NAs.
}
\description{
Calculate the minimum over a rolling window
}
\examples{
min <- minOver(sample_ohlc_data$Close, 9L)
}
