% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{more}
\alias{more}
\title{Print More Rows of Ichimoku Objects}
\usage{
more(rows)
}
\arguments{
\item{rows}{(optional) specify the number of rows to print; defaults to all
rows if not supplied or non-numeric.}
}
\value{
The ichimoku object contained in \code{\link{.Last.value}} (invisibly)
    or else invisible NULL (if .Last.value is not an ichimoku object).
    The ichimoku object data is printed to the console.
}
\description{
After calling or invoking the default print method for ichimoku objects, the
    console output will display \code{-- omitted x rows} if the entire data
    does not fit on-screen. Use \code{more()} to display more rows.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
cloud
more(25)
more()

}
