% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{oanda_switch}
\alias{oanda_switch}
\title{Switch Default OANDA Server}
\usage{
oanda_switch()
}
\value{
Invisible NULL. A message informs the resulting default server setting.
}
\description{
Switch the default OANDA fxTrade server from 'practice' to 'live' or vice versa.
    Settings persist for the current session only.
}
\details{
The default server at the start of a new session is the practice server.
    Call this function to switch to the live server.

    This function can be used to toggle between the practice and live servers.
    Any cached variables for API key, account or instruments list are cleared
    each time this function is called.

    For further details please refer to the OANDA fxTrade API vignette by
    running: \code{vignette("xoanda", package = "ichimoku")}.
}
\examples{
oanda_switch()
oanda_switch()

}
