% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{readArchive}
\alias{readArchive}
\title{Read ichimoku objects from Arrow Archive}
\usage{
readArchive(filename)
}
\arguments{
\item{filename}{string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
The object that was originally archived.
}
\description{
Internal function used to read ichimoku, data.frame and ArrowTabular objects
    from archive storage in the Apache Arrow IPC file format.
}
\keyword{internal}
