% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{index}
\alias{coredata}
\alias{xts}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{xts}{\code{\link[xts]{xts}}}

  \item{zoo}{\code{\link[zoo]{coredata}}, \code{\link[zoo]{index}}}
}}

