% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku.R
\name{rplot}
\alias{rplot}
\title{rplot Reactive Ichimoku Cloud Plot}
\usage{
rplot(
  x,
  ticker,
  theme = c("original", "dark", "solarized", "mono"),
  message,
  strat = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class 'ichimoku'.}

\item{ticker}{(optional) specify a ticker (or other text) to include in the
chart heading. If not set, the ticker saved within the ichimoku object
will be used.}

\item{theme}{[default 'original'] with alternative choices of 'dark',
'solarized' or 'mono'.}

\item{message}{(optional) specify a chart message to display under the title.}

\item{strat}{[default TRUE] if the ichimoku object contains a strategy, the
periods for which the strategy results in a position will be shaded. Set
to FALSE to turn off this behaviour.}

\item{...}{additional arguments to be passed along.}
}
\value{
Returns a Shiny app object with class 'shiny.appobj'.
}
\description{
Plot Ichimoku Kinko Hyo cloud charts from ichimoku objects in a Shiny app,
    allowing full customisation of chart elements in an interactive environment.
    Intuitive cursor tooltip allows ready access to the data directly from the
    chart.
}
\details{
This function has a dependency on the 'shiny' package.
}
\examples{
if(interactive()) {
# Only run examples in interactive R sessions
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
rplot(cloud)
}

}
