% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimdf}
\alias{trimdf}
\title{Trim NA Values}
\usage{
trimdf(x)
}
\arguments{
\item{x}{the data.frame to trim.}
}
\value{
The data.frame 'x' with NA values removed.
}
\description{
Trim NA values from a data.frame. A streamlined version of na.omit.data.frame.
}
\examples{
data <- data.frame(c(1:4, NA), c(NA, 2:5))
trimdf(data)

}
