% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{writeArchive}
\alias{writeArchive}
\title{Write ichimoku objects to Arrow Archive}
\usage{
writeArchive(object, filename)
}
\arguments{
\item{object}{an ichimoku object or dataframe returned by \code{\link{oanda}}.}

\item{filename}{string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
Invisible NULL. 'object' is written to 'file' as a side effect.
}
\description{
Internal function used to write ichimoku objects to archive storage in the
    Apache Arrow IPC file format.
}
\keyword{internal}
