% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{writeMatrix}
\alias{writeMatrix}
\title{writeMatrix}
\usage{
writeMatrix(x, pairs, p2, xlen, type)
}
\arguments{
\item{x}{an ichimoku object or coredata matrix of an ichimoku object.}

\item{pairs}{a 2-column matrix of pairs of column names.}

\item{p2}{length of second ichimoku period.}

\item{xlen}{number of rows of 'x'.}

\item{type}{type of matrix to return, 'boolean' or 'numeric'.}
}
\value{
A matrix with column headings.
}
\description{
Internal function used by mlgrid to create matrices of ichimoku representations.
}
\keyword{internal}
