% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Read/write Objects <> Archive Files with Data Verification}
\usage{
archive(..., object, file)
}
\arguments{
\item{...}{unnamed arguments will be parsed as 'file' if there is only one
argument, 'object' and 'file' if there are two arguments.}

\item{object}{(for write operations) an object.}

\item{file}{the name of the file or a connection where the object is saved to
or read from.}
}
\value{
For read operations: the object originally archived.

    For write operations: invisible NULL. 'object' is written to 'file'.
}
\description{
Read and write objects to/from archival storage in the native RData format,
    with verification of data integrity.
}
\details{
For read operations: specify only 'file'. 'file' is read and the
    return value may be assigned to an object. A confirmation message is
    issued if the file read operation has been successful.

    For write operations: specify both 'object' and 'file'. 'object' will be
    written to 'file'. A confirmation message is issued if the file write
    operation has been successful.
}
\section{Data Verification}{


    If the 'openssl' package is available, a sha256 hash of the original
    object is written to the archive. This allows the data integrity of the
    restored object to be verified when the archive is read back.

    For write operations: confirmation of the sha256 hash is displayed if
    this has been successfully written to file.

    For read operations: a 'data verified' message is issued if the sha256
    hash found within the data file has been authenticated.
}

\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
file <- tempfile()

archive(cloud, file)

restored <- archive(file)

unlink(file)

}
