% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{archive}
\alias{archive}
\title{Read/write ichimoku objects <> Arrow Archive}
\usage{
archive(..., object, filename)
}
\arguments{
\item{...}{additional parameters not used by this function.}

\item{object}{(for write operations) an ichimoku object.}

\item{filename}{string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
For read operations: the ichimoku object originally archived.

    For write operations: invisible NULL. 'object' is written to
    'filename' as a side effect.
}
\description{
Used to read and write ichimoku objects to/from archival storage in the
    Apache Arrow IPC file format.
}
\details{
For read operations, please specify only 'filename'. 'filename' is
    read and the return value may be assigned to an object.

    For write operations: please specify both 'object' and 'filename'.
    'object' will be written to 'filename'. Confirmation is printed to the
    console if the file write operation has been successful.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

filename <- tempfile()
archive(cloud, filename)

restored <- archive(filename)
identical(cloud, restored)

unlink(filename)

}
