% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Read/write objects <> Arrow Archive}
\usage{
archive(..., object, filename)
}
\arguments{
\item{...}{unnamed arguments will be parsed as 'filename' if there is only
one argument, 'object' and 'filename' if there are two arguments.}

\item{object}{(for write operations) an 'ichimoku', 'data.frame' or
'ArrowTabular' object.}

\item{filename}{string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
For read operations: the object originally archived.

    For write operations: invisible NULL. 'object' is written to 'filename'
    as a side effect.
}
\description{
Used to read and write ichimoku, data.frame and ArrowTabular objects to/from
    archival storage in the Apache Arrow IPC file format.
}
\details{
For read operations, please specify only 'filename'. 'filename' is
    read and the return value may be assigned to an object.

    For write operations: please specify both 'object' and 'filename'.
    'object' will be written to 'filename'. Confirmation is printed to the
    console if the file write operation has been successful.
}
\section{Further Details}{


    This function requires the 'arrow' package to be installed.

    If the 'openssl' package is available, a sha256 hash of the original
    object is written to the archive. This allows the data integrity of the
    restored object to be verified when the archive is read back.
}

\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
filename <- tempfile()

archive(cloud, filename)

restored <- archive(filename)

unlink(filename)

}
