% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{autostrat}
\alias{autostrat}
\title{Automated Ichimoku Strategies}
\usage{
autostrat(x, n = 8, dir = c("long", "short"), level = 1)
}
\arguments{
\item{x}{an ichimoku object.}

\item{n}{[default 8] select top n number of strategies to return.}

\item{dir}{[default 'long'] trade direction, either 'long' or 'short'.}

\item{level}{[default 1] set to 2 to also return combined strategies.}
}
\value{
A list of 'n' ichimoku objects containing strategies. The
    cumulative log returns for all strategies as well as the summaries for
    the 'n' top strategies are saved as attributes to the list. The strategy
    summaries are printed to the console as a side effect.
}
\description{
Generate a list of the top performing ichimoku cloud strategies based on
    indicator conditions of the form 'c1 > c2' or level 2 strategies based on
    combined indicator conditions of the form 'c1 > c2 & c3 > c4'.
}
\details{
Ichimoku objects for each strategy are returned as a list. The
    cumulative log returns for all strategies as well as the summaries for
    the 'n' top strategies are saved as attributes to the list. This
    information may be viewed by using look() on the returned list.

    Each individual ichimoku object may be accessed via its position in the
    list, e.g. [[1]] for the 1st item, or by using \code{\link{look}}
    specifying the parameter 'which'.
}
\section{Further Details}{

    Please refer to the strategies vignette by running:
    \code{vignette("strategies", package = "ichimoku")}
}

\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

stratlist <- autostrat(cloud, n = 3)
look(stratlist)
summary(look(stratlist, which = 1))

autostrat(cloud, n = 1, dir = "short", level = 2)

}
