% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_POSIXct}
\alias{format_POSIXct}
\title{Format POSIXct}
\usage{
format_POSIXct(x)
}
\arguments{
\item{x}{an object of class \sQuote{POSIXct}.}
}
\value{
A character string
}
\description{
Converts a POSIXct double timestamp into a character string much faster
than \code{\link{format.POSIXct}}, with the same output as the method
default.
}
\examples{
time <- Sys.time()
format(time)
format_POSIXct(time)

}
