% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku-package.R
\docType{package}
\name{ichimoku-package}
\alias{ichimoku-package}
\title{ichimoku: Compute and Plot Ichimoku Kinko Hyo Cloud Charts}
\description{
An implementation of the 'Ichimoku Kinko Hyo' charting system, also commonly
known as 'cloud charts', providing both publication-ready and fully-interactive
charts for analysis. Originating from and popularised in Japan, the technique
is a refinement on candlestick charting, in widespread use on trading floors
worldwide. Translating to 'one-glance equilibrium chart', it allows the price
action and market structure of financial securities to be determined
'at-a-glance'.
}
\section{Principal ichimoku functions}{

\itemize{
    \item{\code{\link{ichimoku}}} {to create an ichimoku object.}
    \item{\code{\link{plot.ichimoku}}} {to plot a cloud chart from an
    ichimoku object.}
    \item{\code{\link{iplot}}} {to plot an interactive cloud chart from an
    ichimoku object.}
}
}

\references{
Package website:
    \url{https://shikokuchuo.net/ichimoku/}

    The most recent version of the package may be found at
    \url{https://github.com/shikokuchuo/ichimoku/}
}
\author{
Charlie Gao <\email{charlie.gao@shikokuchuo.net}>
}
