% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku-package.R
\docType{package}
\encoding{UTF-8}
\name{ichimoku-package}
\alias{ichimoku-package}
\title{ichimoku: Visualization and Tools for Ichimoku Kinko Hyo Strategies}
\description{
An implementation of 'Ichimoku Kinko Hyo', also commonly known as 'cloud
    charts'. Static and interactive visualizations with tools for creating,
    backtesting and development of quantitative 'ichimoku' strategies. As
    described in Sasaki (1996, ISBN:4925152009), the technique is a
    refinement on candlestick charting originating from Japan, now in
    widespread use in technical analysis worldwide. Translating as
    'one-glance equilibrium chart', it allows the price action and market
    structure of financial securities to be determined 'at-a-glance'.
    Incorporates an interface with the OANDA fxTrade API
    \url{https://developer.oanda.com/} for retrieving historical and live
    streaming price data for major currencies, metals, commodities,
    government bonds and stock indices.
}
\section{Principal ichimoku functions}{

Data
\itemize{
    \item{\code{\link{ichimoku}}} {to create an ichimoku object from price
    data.}
    \item{\code{\link{archive}}} {for reading and writing ichimoku objects
    to/from Apache Arrow archive files.}
    \item{\code{\link{oanda}}} {to retrieve price data from the OANDA fxTrade
    API.}
    \item{\code{\link{oanda_stream}}} {to stream a live data feed from the
    OANDA fxTrade API.}
}
Visualization
\itemize{
    \item{\code{\link{plot.ichimoku}}} {to plot a cloud chart from an
    ichimoku object.}
    \item{\code{\link{iplot}}} {to plot an interactive cloud chart from an
    ichimoku object.}
    \item{\code{\link{oanda_chart}}} {to create live updating ichimoku cloud
    charts using OANDA data.}
    \item{\code{\link{oanda_studio}}} {a complete live analysis environment
    using OANDA data implemented in R Shiny.}
}
Strategies & ML
\itemize{
    \item{\code{\link{strat}}} {to augment an ichimoku object with a strategy,
    including complex combined and asymmetric strategies.}
    \item{\code{\link{stratcombine}}} {to create custom combined strategies.}
    \item{\code{\link{autostrat}}} {to automatically evaluate and rank
    top-performing strategies.}
    \item{\code{\link{mlgrid}}} {to generate a numeric representation of the
    relationship between ichimoku cloud chart elements.}
}
}

\references{
Sasaki, H. 佐々木 英信 (1996), \emph{一目均衡表の研究
    [ichimoku kinkouhyou no kenkyuu]}. Tokyo, Japan: Toushi Radar.

    OANDA' and 'fxTrade' are trademarks owned by OANDA Corporation, an entity
    unaffiliated with the ichimoku package.

    Gao, C. (2021), \emph{ichimoku: Visualization and Tools for Ichimoku
    Kinko Hyo Strategies}. R package version 0.3.51,
    \url{https://CRAN.R-project.org/package=ichimoku}.
}
\author{
Charlie Gao <\email{charlie.gao@shikokuchuo.net}>
}
