% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku.R
\name{is.ichimoku}
\alias{is.ichimoku}
\title{is.ichimoku}
\usage{
is.ichimoku(x)
}
\arguments{
\item{x}{an object.}
}
\value{
A logical value of TRUE if 'x' is of class 'ichimoku', otherwise FALSE.
}
\description{
A function for checking if an object is an ichimoku object.
}
\examples{
\dontrun{
# where 'cloud' is an ichimoku object created by the ichimoku() function:

is.ichimoku(cloud)
# [1] TRUE
}

}
