% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_df}
\alias{matrix_df}
\title{Convert matrix to data.frame}
\usage{
matrix_df(x, keep.attrs)
}
\arguments{
\item{x}{a matrix.}

\item{keep.attrs}{(optional) if set to TRUE, will preserve any custom
attributes set on the original object.}
}
\value{
A 'data.frame' object. If the matrix has row names, these are
    retained by the dataframe.
}
\description{
A performant 'matrix' to 'data.frame' constructor with no data validation or
    checking.
}
\examples{
cloud <- ichimoku(sample_ohlc_data)
mcloud <- as.matrix(cloud)
df <- matrix_df(mcloud)
str(df)
str(rownames(df))

}
