% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxOver}
\alias{maxOver}
\title{Maximum over a rolling window}
\usage{
maxOver(x, window)
}
\arguments{
\item{x}{a vector.}

\item{window}{size of the rolling window.}
}
\value{
A vector of the same length as 'x' with elements 1 to
    (length(window) - 1) containing NAs.
}
\description{
Calculate the maximum over a rolling window
}
\examples{
max <- maxOver(sample_ohlc_data$Close, 9L)
}
