% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{more}
\alias{more}
\title{Print More Rows of Ichimoku Objects}
\usage{
more()
}
\value{
The ichimoku object originally printed (invisibly) or else invisible
    NULL (if the last returned object was not an ichimoku object). The
    ichimoku object data is printed to the console.
}
\description{
After calling or invoking the default print method for ichimoku objects, the
    console output will display \code{# ... with x more rows} if the entire
    data does not fit on-screen. Call \code{more()} to show up to 100 more rows.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

cloud
more()

}
