% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{oanda_switch}
\alias{oanda_switch}
\title{Switch Default OANDA Server}
\usage{
oanda_switch()
}
\value{
Invisible NULL. A message informs the resulting default server setting.
}
\description{
Switch the default OANDA fxTrade server from 'practice' to 'live' or vice versa.
}
\details{
This function is designed to be run once at the start of each session
    if using a live account.

    Re-load the package using \code{library(ichimoku)} before calling
    \code{oanda_switch()} to clear cached account variables if necessary.

    For further details please refer to the OANDA fxTrade API vignette by
    running: \code{vignette("xoanda", package = "ichimoku")}.
}
\examples{
oanda_switch()
oanda_switch()

}
