% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichimoku.R
\name{print.ichimoku}
\alias{print.ichimoku}
\title{print.ichimoku}
\usage{
\method{print}{ichimoku}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class 'ichimoku'.}

\item{plot}{set to TRUE by default. Set to FALSE to prevent automatic
plotting of the ichimoku cloud chart.}

\item{...}{other arguments to be passed along.}
}
\value{
The ichimoku object 'x' passed as parameter.
}
\description{
Custom print method for ichimoku objects.
}
\details{
This function is an S3 method for the generic function print() for
    class 'ichimoku'. It can be invoked by calling print(x) on an object 'x'
    of class 'ichimoku'.
}
\section{Further Details}{

    Please refer to the reference vignette by running:
    \code{vignette("reference", package = "ichimoku")}
}

\examples{
\dontrun{
# where 'cloud' is an ichimoku object created by the ichimoku() function:

print(cloud)
print(cloud, plot = FALSE)
}

}
