% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{readArchive}
\alias{readArchive}
\title{Read Objects from Archive}
\usage{
readArchive(file)
}
\arguments{
\item{file}{the name of the file or a connection where the object is saved to
or read from.}
}
\value{
The object that was originally archived.
}
\description{
Internal function used to read objects from native RData files with stored
    sha256 hash values.
}
\keyword{internal}
