% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{readArchive}
\alias{readArchive}
\title{Read ichimoku objects from Arrow Archive}
\usage{
readArchive(filename)
}
\arguments{
\item{filename}{string file path, URI, or OutputStream, or path in a file
system (SubTreeFileSystem).}
}
\value{
The ichimoku object that was originally archived.
}
\description{
Used to read ichimoku objects from archive storage in the Apache Arrow IPC
    file format.
}
\keyword{internal}
