% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{relative}
\alias{relative}
\title{Relative Numeric Representation}
\usage{
relative(x)
}
\arguments{
\item{x}{an ichimoku object.}
}
\value{
A data frame listing the mean and standard deviation of the numeric
    ichimoku chart representation for the data, the current values for the
    latest available price data, and their relative values in terms of
    standard deviations from the mean. The time index for 'current' is saved
    as an attribute and also printed to the console.
}
\description{
Compare the current numeric representation of the ichimoku cloud chart
    relative to historical values. Aids in determining whether current trading
    behaviour is exceptional or falls within normal ranges.
}
\details{
'relative' is calculated as (current - mean) / sd and represents a
    centred and scaled measure of deviation. The elements are ordered by
    their 'relative' value, so that those with the highest (postive or
    negative) relative deviation are listed first.

    A large 'relative' value indicates exceptional behaviour as opposed to
    being within normal ranges.
}
\section{Further Details}{

    Please refer to the strategies vignette by calling:
    \code{vignette("strategies", package = "ichimoku")}
}

\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
relative(cloud)

}
