% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{relative}
\alias{relative}
\title{Relative Numeric Representation}
\usage{
relative(x)
}
\arguments{
\item{x}{an ichimoku object.}
}
\value{
A data frame containing a statistical summary of the latest ichimoku
    cloud chart representation and how it relates to the historical data.
    The elements are ordered by their absolute 'r value', so that those with
    the highest relative deviation are listed first. The time index value for
    'latest' and the number of observations are printed to the console.
}
\description{
Produce a statistical summary of the latest numeric representation of the
    ichimoku cloud chart relative to historical values. Aids in determining
    if current trading falls within normal or exceptional ranges.
}
\details{
Column 'mean' is the mean value, 'sd' the standard deviation,
    and 'latest' the latest observed values from the data.

    The 'relative' or 'r value' is calculated as (latest - mean) / sd and
    represents a standardised (centred and scaled) measure of deviation.

    Column 'p >= |r|' contains the empirical probability of the observed
    absolute 'r value' or greater.

    Column 'E(|r|)|p' contains the expected absolute 'r value', conditional
    upon being greater than equal to the observed absolute 'r value'.
}
\section{Further Details}{

    Please refer to the strategies vignette by calling:
    \code{vignette("strategies", package = "ichimoku")}
}

\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
relative(cloud)

}
