% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ohlc_data}
\alias{sample_ohlc_data}
\title{Sample OHLC Price Data}
\format{
A data frame with 256 observations of 6 variables:
\itemize{
  \item{time}{ - timestamp of observation [POSIXct]}
  \item{open}{ - opening price [numeric]}
  \item{low}{ - low price [numeric]}
  \item{high}{ - high price [numeric]}
  \item{close}{ - closing price [numeric]}
  \item{volume}{ - volume [integer]}
}
}
\source{
Not applicable: simulated data
}
\usage{
sample_ohlc_data
}
\description{
Simulated prices for a hypothetical financial asset. Created for the purpose
    of demonstrating package functions in examples and vignettes only.
}
\examples{
head(sample_ohlc_data)

}
\keyword{datasets}
