% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ohlc_data}
\alias{sample_ohlc_data}
\title{Sample OHLC Price Data}
\format{
A data frame with 234 observations of 5 variables:
\itemize{
  \item{Date}{ - date of observation}
  \item{Open}{ - daily opening price}
  \item{Low}{ - daily low price}
  \item{High}{ - daily high price}
  \item{Close}{ - daily closing price}
}
}
\source{
Not applicable: simulated data
}
\usage{
sample_ohlc_data
}
\description{
Simulated prices for a hypothetical financial asset. Created for the purpose
    of demonstrating package functions in examples and vignettes only.
}
\examples{
cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

}
\keyword{datasets}
