% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tradingDays}
\alias{tradingDays}
\title{Check if Dates are Trading Days}
\usage{
tradingDays(x, holidays, ...)
}
\arguments{
\item{x}{a POSIXct date object, or vector of POSIXct date objects.}

\item{holidays}{(optional) a vector, or function which outputs a
vector, of dates defined as holidays. If not specified, New Year's and
Christmas day are defined as holidays by default.}

\item{...}{other arguments to be passed along.}
}
\value{
A vector of logical values: TRUE if the corresponding element of 'x'
    is a weekday and not a holiday, otherwise FALSE.
}
\description{
A function for checking if specified dates are trading days.
}
\examples{
tradingDays(sample_ohlc_data$Date[1:7])

}
