% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{writeArchive}
\alias{writeArchive}
\title{Write Objects to Archive}
\usage{
writeArchive(object, file)
}
\arguments{
\item{object}{an object.}

\item{file}{the name of the file or a connection where the object is saved to
or read from.}
}
\value{
Invisible NULL. 'object' is written to 'file' as a side effect.
}
\description{
Internal function used to write objects, along with their sha256 hash value,
    to archive files in the native RData format.
}
\keyword{internal}
