% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{writeStrat}
\alias{writeStrat}
\title{writeStrat}
\usage{
writeStrat(x, strategy, dir)
}
\arguments{
\item{x}{an ichimoku object augmented with strategy columns.}

\item{strategy}{string describing strategy rule.}

\item{dir}{trade direction, either 'long' or 'short'.}
}
\value{
An ichimoku object with the strategy summary set as the attribute
    'strat'.
}
\description{
Internal function used by ichimoku to write strategy summaries to ichimoku
    objects.
}
\details{
The stategy summary may subsequently be accessed by the summary()
    function or via \code{\link{look}}.
}
\keyword{internal}
